pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["Sundance"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.3.28"
supabase = "^2.15.0"
python-dotenv = "^1.1.0"
google-generativeai = "^0.8.4"
langchain-anthropic = "^0.3.10"
pydot = "^3.0.4"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
